# src/campaign/swagger/swagger.yaml
consumes:
- application/json
- application/xml
- application/gob
- application/x-gob
definitions:
...
  CampaignCollection:
    description: CampaignCollection is the media type for an array of Campaign (default
      view)
    example:
    - activeHours: 1.14478196e+18
      activeStartHour: 6
      activeStartMinute: 0
      approvedBy: 8.968521e+18
      approvedOn: 4.198896e+18
      campaignId: Sit non voluptatibus officia.
      currentExecutionCycleStartTime: 2.2921524e+18
      currentTargetVolume: 5.666694e+17
      endDate: 6.233304e+18
      executionFrequency: 1
      lastExecutionTime: 6.822325e+18
      pollingInterval: 1.3490164e+18
      productId: 5oz5fb
      startDate: 2.3695352e+18
      status: 1.1657216e+18
    - activeHours: 1.14478196e+18
      activeStartHour: 6
      activeStartMinute: 0
      approvedBy: 8.968521e+18
      approvedOn: 4.198896e+18
      campaignId: Sit non voluptatibus officia.
      currentExecutionCycleStartTime: 2.2921524e+18
      currentTargetVolume: 5.666694e+17
      endDate: 6.233304e+18
      executionFrequency: 1
      lastExecutionTime: 6.822325e+18
      pollingInterval: 1.3490164e+18
      productId: 5oz5fb
      startDate: 2.3695352e+18
      status: 1.1657216e+18
    - activeHours: 1.14478196e+18
      activeStartHour: 6
      activeStartMinute: 0
      approvedBy: 8.968521e+18
      approvedOn: 4.198896e+18
      campaignId: Sit non voluptatibus officia.
      currentExecutionCycleStartTime: 2.2921524e+18
      currentTargetVolume: 5.666694e+17
      endDate: 6.233304e+18
      executionFrequency: 1
      lastExecutionTime: 6.822325e+18
      pollingInterval: 1.3490164e+18
      productId: 5oz5fb
      startDate: 2.3695352e+18
      status: 1.1657216e+18
    items:
      $ref: '#/definitions/Campaign'
    title: 'Mediatype identifier: application/ts.campaign; type=collection; view=default'
    type: array
...
  CampaignExecutionContext:
    description: CampaignExecutionContext media type (default view)
    example:
      campaignId: Et odio atque reiciendis eaque quam dignissimos.
      endTime: 1972-11-24T20:19:43Z
      executionId: Sequi saepe reprehenderit.
      numMessagesSent: 0.9511033
      startTime: 1998-11-27T21:02:12Z
    properties:
      campaignId:
        description: campaign id
        example: Et odio atque reiciendis eaque quam dignissimos.
        type: string
      endTime:
        description: execution endTime time
        example: 1972-11-24T20:19:43Z
        format: date-time
        type: string
      executionId:
        description: execution id
        example: Sequi saepe reprehenderit.
        type: string
      numMessagesSent:
        description: Number of message sent in this execution
        example: 0.9511033
        format: double
        type: number
      startTime:
        description: execution start time
        example: 1998-11-27T21:02:12Z
        format: date-time
        type: string
    title: 'Mediatype identifier: application/ts.campaignexecution; view=default'
    type: object
...
  leadPoolLength:
    description: leadPoolLength media type (default view)
    example:
      length: Sequi voluptas expedita sint at.
    properties:
      length:
        description: length of lead pool
        example: Sequi voluptas expedita sint at.
        type: string
    title: 'Mediatype identifier: application/ts.leadpool; view=default'
    type: object
...
host: localhost:8050
info:
  description: Trusting social campaign management
  title: The Campaign Management API
  version: ""
paths:
  /campaigns/:
    get:
      description: Returns all the campaigns
      operationId: campaigns#getAll
      parameters:
      - description: The state
        in: query
        name: state
        required: false
        type: number
      produces:
      - application/text
      - application/ts.campaign; type=collection
      responses:
        "200":
          description: This is the success response.
          schema:
            $ref: '#/definitions/CampaignCollection'
        "400":
          description: Response sent for bad requests.
        "401":
          description: Response sent for unauthorized requests.
        "500":
          description: Response sent for Database or Internal Server Errors.
      schemes:
      - http
      summary: getAll campaigns
      tags:
      - campaigns
    post:
      description: Creates a campaign.
      operationId: campaigns#create
      parameters:
      - description: The Campaign object
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/CampaignPayload'
      produces:
      - application/text
      - application/ts.campaign
      responses:
        "201":
          description: campaign created successfully.
          schema:
            $ref: '#/definitions/Campaign'
        "400":
          description: Response sent for bad requests.
        "401":
          description: Response sent for unauthorized requests.
        "500":
          description: Response sent for Database or Internal Server Errors.
      schemes:
      - http
      summary: create campaigns
      tags:
      - campaigns
  /campaigns/{campaignId}:
    delete:
      description: Deletes a terminated campaign.
      operationId: campaigns#delete
      parameters:
      - description: The id of the campaign to be updated
        in: path
        name: campaignId
        required: true
        type: string
      produces:
      - application/text
      responses:
        "204":
          description: This is the success response.
        "400":
          description: Response sent for bad requests.
        "401":
          description: Response sent for unauthorized requests.
        "404":
          description: Item not found
        "500":
          description: Response sent for Database or Internal Server Errors.
      schemes:
      - http
      summary: delete campaigns
      tags:
      - campaigns
    get:
      description: Returns a campaign with  specific id.
      operationId: campaigns#get
      parameters:
      - description: The campaignId
        in: path
        name: campaignId
        required: true
        type: string
      produces:
      - application/text
      - application/ts.campaign
      responses:
        "200":
          description: This is the success response.
          schema:
            $ref: '#/definitions/CampaignDetailed'
        "400":
          description: Response sent for bad requests.
        "401":
          description: Response sent for unauthorized requests.
        "404":
          description: Response when Not Found
        "500":
          description: Response sent for Database or Internal Server Errors.
      schemes:
      - http
      summary: get campaigns
      tags:
      - campaigns
    patch:
      description: Updates a campaign.
      operationId: campaigns#update
      parameters:
      - description: The id of the campaign to be updated
        in: path
        name: campaignId
        required: true
        type: string
      - description: The Campaign object
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/CampaignUpdatePayload'
      produces:
      - application/text
      - application/ts.campaign
      responses:
        "202":
          description: This is the success response.
          schema:
            $ref: '#/definitions/Campaign'
        "400":
          description: Response sent for bad requests.
        "401":
          description: Response sent for unauthorized requests.
        "404":
          description: Item not found
        "500":
          description: Response sent for Database or Internal Server Errors.
      schemes:
      - http
      summary: update campaigns
      tags:
      - campaigns
  /campaigns/{campaignId}/executions:
    get:
      description: Returns all campaign  execution details.
      operationId: campaigns#getAllCampaignExecution
      parameters:
      - description: The campaignId
        in: path
        name: campaignId
        required: true
        type: string
      produces:
      - application/text
      - application/ts.campaignexecution
      responses:
        "200":
          description: This is the success response.
          schema:
            $ref: '#/definitions/CampaignExecutionContext'
        "400":
          description: Response sent for bad requests.
        "401":
          description: Response sent for unauthorized requests.
        "404":
          description: Response for Not Found
        "500":
          description: Response sent for Database or Internal Server Errors.
      schemes:
      - http
      summary: getAllCampaignExecution campaigns
      tags:
      - campaigns
  /campaigns/{campaignId}/executions/{executionId}:
    get:
      description: Returns a specific campaign execution details
      operationId: campaigns#getCampaignExecution
      parameters:
      - description: The campaign Id
        in: path
        name: campaignId
        required: true
        type: string
      - description: The execution Id
        in: path
        name: executionId
        required: true
        type: string
      produces:
      - application/text
      - application/ts.campaignexecution
      responses:
        "200":
          description: This is the success response.
          schema:
            $ref: '#/definitions/CampaignExecutionContext'
        "400":
          description: Response sent for bad requests.
        "401":
          description: Response sent for unauthorized requests.
        "404":
          description: Response for Not Found
        "500":
          description: Response sent for Database or Internal Server Errors.
      schemes:
      - http
      summary: getCampaignExecution campaigns
      tags:
      - campaigns
  /lead/{productId}:
    get:
      operationId: lead#get
      parameters:
      - description: the product id
        in: path
        name: productId
        required: true
        type: string
      produces:
      - application/ts.leadpool
      responses:
        "201":
          description: The length of lead pool
          schema:
            $ref: '#/definitions/leadPoolLength'
        "404":
          description: length not found
      schemes:
      - http
      summary: get lead
      tags:
      - lead
  /messagecontents/:
    get:
      description: Returns all the messages.
      operationId: messagecontents#list
      produces:
      - application/text
      - application/ts.campaign.messagecontent
      responses:
        "200":
          description: This is the success response.
          schema:
            $ref: '#/definitions/CampaignMessageContent'
        "400":
          description: Response sent for bad requests.
        "401":
          description: Response sent for unauthorized requests.
        "500":
          description: Response sent for Database or Internal Server Errors.
      schemes:
      - http
      summary: list messagecontents
      tags:
      - messagecontents
    post:
      description: Creates a message content.
      operationId: messagecontents#create
      parameters:
      - description: The Message content object
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/MessageContentPayload'
      produces:
      - application/text
      - application/ts.campaign.messagecontent
      responses:
        "201":
          description: Message content  created successfully.
          schema:
            $ref: '#/definitions/CampaignMessageContent'
        "400":
          description: Response sent for bad requests.
        "401":
          description: Response sent for unauthorized requests.
        "500":
          description: Response sent for Database or Internal Server Errors.
      schemes:
      - http
      summary: create messagecontents
      tags:
      - messagecontents
  /messagecontents/{messageId}:
    delete:
      description: Deletes a message.
      operationId: messagecontents#delete
      parameters:
      - description: The message content id
        in: path
        name: messageId
        required: true
        type: string
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/MessageContentDeletePayload'
      produces:
      - application/text
      responses:
        "204":
          description: Deleted successfully
        "400":
          description: Response sent for bad requests.
        "401":
          description: Response sent for unauthorized requests.
        "404":
          description: Item not found
        "500":
          description: Response sent for Database or Internal Server Errors.
      schemes:
      - http
      summary: delete messagecontents
      tags:
      - messagecontents
    get:
      description: Returns the messages defined for specific product type.
      operationId: messagecontents#get
      parameters:
      - description: The message contentid
        in: path
        name: messageId
        required: true
        type: string
      produces:
      - application/text
      - application/ts.campaign.messagecontent
      responses:
        "200":
          description: This is the success response.
          schema:
            $ref: '#/definitions/CampaignMessageContent'
        "400":
          description: Response sent for bad requests.
        "401":
          description: Response sent for unauthorized requests.
        "404":
          description: Response for Not Found
        "500":
          description: Response sent for Database or Internal Server Errors.
      schemes:
      - http
      summary: get messagecontents
      tags:
      - messagecontents
    put:
      description: Updates a  message.
      operationId: messagecontents#update
      parameters:
      - in: path
        name: messageId
        required: true
        type: string
      - description: The Message content object
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/MessageContentUpdatePayload'
      produces:
      - application/text
      - application/ts.campaign.messagecontent
      responses:
        "202":
          description: This is the success response.
          schema:
            $ref: '#/definitions/CampaignMessageContent'
        "400":
          description: Response sent for bad requests.
        "401":
          description: Response sent for unauthorized requests.
        "404":
          description: Item not found
        "500":
          description: Response sent for Database or Internal Server Errors.
      schemes:
      - http
      summary: update messagecontents
      tags:
      - messagecontents
  /products/:
    post:
      description: creates a product
      operationId: products#create
      parameters:
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/Product Payload'
      responses:
        "201":
          description: Product Created Successfully
      schemes:
      - http
      summary: create products
      tags:
      - products
  /products/{productId}:
    get:
      operationId: products#get
      parameters:
      - description: the product id
        in: path
        name: productId
        required: true
        type: string
      produces:
      - application/ts.campaign.product
      responses:
        "201":
          description: get product details by id
          schema:
            $ref: '#/definitions/campaignProductMedia'
        "404":
          description: product detail not found
      schemes:
      - http
      summary: get products
      tags:
      - products
  /smstrackerservice/:
    post:
      description: To send an sms
      operationId: smstrackerservice#create
      parameters:
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/smsPayload'
      produces:
      - application/ts.smstracker
      responses:
        "200":
          description: 'sms created '
          schema:
            $ref: '#/definitions/smsMedia'
      schemes:
      - http
      summary: create smstrackerservice
      tags:
      - smstrackerservice
...
produces:
- application/json
- application/xml
- application/gob
- application/x-gob
schemes:
- http
swagger: "2.0"